Apply two patch files that were submitted by Dirk Lembke (thank you Dirk) The changes of patch 1 are:
- Make source entries in OpenCms library editable
- Add source location to preferences
- Set source location for opencms jars automatically (opencms.* and
org.opencms.*) to value in preferences
- order library entries by name (easier to find an entry)

With these changes you can directly debug opencms or jump to sources by
"Open Declaration". 

I have the problem that it's not possible to build your plugin with the
unchanged sources from repository and a new eclipse installation.
Debugging is possible but building with the export wizard ends with an
error.
Patch 2 should resolve this problem.